requirements.txt
Flask==2.0.2
SQLAlchemy==1.4.27
psycopg2-binary==2.9.3
pymongo==3.12.1
redis==4.0.2
elasticsearch==7.15.1
requests==2.26.0
gunicorn==20.1.0